c_image.py: use cc instead of gcc
Compile using ``cc`` instead of ``gcc`` to support systems where the
default C compiler is not GCC.  For example, GCC does not ship in
FreeBSD 10, but Clang is installed as ``cc`` and works fine.  On
systems where GCC is the default C compiler, it is usually installed
as ``cc`` as well, so this change should not result in problems for
existing users.