Correction on comment-detection regexp
During the config parse, comments are removed using a complex regexp,
matching (among other things) '//'. A problem arises when your
configuration contains a link of some sort (such as 'http://' ...). The
regexp then consider the '//' and what follows as a comment (and your
configuration is consequently considered corrupt).

Escapinf the '/' works for a time, but it fails as soon as you use the
'config' command at runtime, because the configuration is then
rewritten.

Therefore I changed the regexp used to match comments. It's a fairly
complex regexp though, and I'm not 100% positive my change is perfect.
Fell free to review it.